requirements-dev.txt
black==21.7b0
eth-brownie>=1.16.0,<2.0.0
eth_abi==3.0.0